Refactoring of config management
We have now an asset class that takes care of downloading assets (like
neo-extensions.el) using sha256 to decide whether a newer version
exists and should be downloaded.

The same also handles things like initial-frame-properties that is
actually generated by Emacs itself and will be later integrated with
hooks run when the theme or the frame geometry change.

GitOrigin-RevId: 312efc7460920c4f9a5619a117c10f9857b6e89d