✨ feat(neo): Core neo (#31)
* ✨ feat(neo): Core neo

This is just enough neo to load extensions. As of now we don't have
any so this PR might not be self-contained and might die an horrendous
death trying to accedd things that are not there.

* 🧪 test(neo): try to fix github tests for emacs

* 🧪 test(neo): add loading elisp files

* 🧪 test(neo): don't do tests for now, they don't work

* 🧪 test(neo): don't run bazel build tests in neo

For now we don't have anything built. This will change when we publish
with copybara, but even then it probably doesn't make sense to build
the entire Omega monorepo.

* 🧪 test(neo): paths-filter v4 doesn't exist

GitOrigin-RevId: e91decd4eccd6785a41ff12e59b0cca8a3727e7e